requirements.txt
flask==2.3.2
slack-sdk==3.36.0
python-dotenv==1.1.1
